Refactor script.js
