+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Expenses Tracker</title>
     <link rel="stylesheet" href="./styles/index.css" />
-    <script src="./script.js" defer></script>
+    <script src="./js/index.js" type="module" defer></script>
   </head>
   <body>
     <header class="header">
