--- script.js
@@ -1,162 +1,0 @@
-class Expenses {
-  static create(data) {
-    localStorage.setItem('expenses', JSON.stringify([...this.read(), data]));
-  }
-
-  static read() {
-    if (!localStorage.getItem('expenses')) {
-      return [];
-    }
-
-    return JSON.parse(localStorage.getItem('expenses'));
-  }
-
-  static update(data, id) {
-    const expenses = this.read();
-    expenses.splice(id, 1, data);
-    localStorage.setItem('expenses', JSON.stringify(expenses));
-  }
-
-  static remove(id) {
-    const expenses = this.read();
-    expenses.splice(id, 1);
-    localStorage.setItem('expenses', JSON.stringify(expenses));
-  }
-
-  static empty() {
-    localStorage.removeItem('expenses');
-  }
-
-  static get balance() {
-    if (!this.read().length) {
-      return 0;
-    }
-
-    return this.read()
-      .map((obj) => +obj.amt)
-      .reduce((p, c) => p + c);
-  }
-}
-
-const formEl = document.querySelector('.form');
-const formBtn = document.querySelectorAll('.form button');
-const inputEls = document.querySelectorAll('.form input');
-
-render();
-resetForm();
-
-formBtn[2].addEventListener('click', () => {
-  const data = {};
-
-  inputEls.forEach((el) => {
-    data[el.id] = el.value;
-  });
-
-  const action = formEl.getAttribute('data-action');
-  const index = formEl.getAttribute('data-id');
-  let isFormValid = true;
-
-  for (const key in data) {
-    const value = data[key];
-
-    if (!value.trim()) {
-      [...inputEls].find((el) => el.id === key).classList.add('error');
-      isFormValid = false;
-    } else {
-      [...inputEls].find((el) => el.id === key).classList.remove('error');
-    }
-  }
-
-  if (isFormValid) {
-    if (action === 'create') Expenses.create(data);
-    else if (action === 'update') Expenses.update(data, index);
-
-    resetForm();
-    render();
-  }
-});
-
-formBtn[1].addEventListener('click', () => {
-  const index = formEl.getAttribute('data-id');
-  Expenses.remove(index);
-  resetForm();
-  render();
-});
-
-formBtn[0].addEventListener('click', resetForm);
-
-function render() {
-  const historyEl = document.querySelector('.history');
-  historyEl.innerHTML = '';
-
-  historyEl.innerHTML += `
-    <button class="empty-expenses" onclick="emptyExpenses()">🚮</button>
-    <div class="balance">
-      <span>Balance :</span>
-      <span class="${Expenses.balance >= 0 ? 'positive' : 'negative'} amt">
-        ${formatMoney(Expenses.balance)}
-      </span>
-    </div>
-  `;
-
-  historyEl.innerHTML += Expenses.read()
-    .map(
-      (obj, i) =>
-        `<div class="transaction" onclick="update(${i})">
-          <span class="date">${obj.date}</span>
-          <span class="desc">${obj.desc}</span>
-          <span class="amt ${obj.amt >= 0 ? 'positive' : 'negative'}">
-            ${formatMoney(obj.amt)}
-          </span>
-        </div>`
-    )
-    .reverse()
-    .join('');
-}
-
-function update(index) {
-  formEl.setAttribute('data-action', 'update');
-  formEl.setAttribute('data-id', index);
-  formBtn[2].textContent = 'Update';
-
-  inputEls.forEach((el) => {
-    el.value = Expenses.read()[index][el.id];
-  });
-}
-
-function emptyExpenses() {
-  Expenses.empty();
-  render();
-}
-
-function resetForm() {
-  formEl.setAttribute('data-action', 'create');
-  formEl.removeAttribute('data-id');
-  formBtn[2].textContent = 'Create';
-
-  inputEls.forEach((el) => {
-    el.value = '';
-    el.classList.remove('error');
-  });
-}
-
-function formatMoney(money) {
-  let result = String(+money).split('.');
-  const decimalNumber = result.splice(1, 1)[0];
-  result = result[0].split('').reverse();
-
-  const len = result.length;
-
-  for (let x = 3; x < len; x += 3) {
-    if (result[x] === '-') continue;
-    result.splice(x, 1, `${result[x]}, `);
-  }
-
-  return decimalNumber
-    ? `${result.reverse().join('')}.${decimalNumber}`
-    : result.reverse().join('');
-}
-
-function isEven(num) {
-  return num % 2 ? false : true;
-}
